third_party/WebKit/LayoutTests/fast/forms/time-multiple-fields/time-multiple-fields-ax-value-changed-notification.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<input id=test type="time" value="12:34:56.789">
<script>
description('This test checks value changed accessibility notifications.');

function enableAccessibility()
{
    if (!window.accessibilityController) {
        debug("Please run inside DumpRenderTree.");
        return;
    }

    accessibilityController.addNotificationListener(function (element, notification) {
        if (notification == 'Focus' || notification == 'ValueChanged') {
            debug(notification + ' ' + element.description + '=' + element.valueDescription);
        }
    });
}

function keyDown(key, modifiers)
{
    if (!window.eventSender)
        return;
    eventSender.keyDown(key, modifiers);
}

var testInput = document.getElementById("test");

enableAccessibility();
testInput.focus();
keyDown('\t');
keyDown('4');
keyDown('ArrowUp');

window.jsTestIsAsync = true;

window.setTimeout(function () {
    debug('');
    testInput.parentNode.removeChild(testInput);
    finishJSTest();
}, 1);
</script>
</body>
</html>
