third_party/WebKit/LayoutTests/fast/forms/time-multiple-fields/time-multiple-fields-ax-value-changed-notification-expected.txt

This test checks value changed accessibility notifications.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Focus Hours=AXValueDescription: 12
Focus Minutes=AXValueDescription: 34
ValueChanged Minutes=AXValueDescription: 05
ValueChanged Minutes=AXValueDescription: 05
ValueChanged Minutes=AXValueDescription: 05
ValueChanged Minutes=AXValueDescription: 05

PASS successfullyParsed is true

TEST COMPLETE
